company_agent/requirements.txt
google-generativeai>=0.8.0
groq>=0.11.0
requests>=2.31.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
langgraph>=0.0.60
langchain-core>=0.2.0
langchain-groq>=0.1.0
langchain-mistralai>=0.1.0
langchain-openai>=0.1.0
pydantic>=2.0.0
